+++ src/utilities.ts
@@ -2,6 +2,7 @@ import * as fs from 'fs';
 import * as path from 'path';
 import * as vscode from 'vscode';
 import * as URL from 'url';
+import { isNullOrUndefined } from 'util';
 
 export class Utilities{
 
@@ -21,7 +22,9 @@ export class Utilities{
 
     public static IsValidUrl(url: string) : boolean {
         try {
-            URL.parse(url);
+            if(url.length < 3) { return false };
+            var urlparsed = URL.parse(url);
+            if( isNullOrUndefined(urlparsed.hostname)) { return false; }
             return true;
           } catch (err) {
             return false;